REFACTOR/ Refactor all project Refactor app, tests, auth, security and settings files
