+++ fast_zero/app.py
@@ -1,103 +1,13 @@
-from fastapi import Depends, FastAPI, HTTPException
-from fastapi.security import OAuth2PasswordRequestForm
-from sqlalchemy import select
-from sqlalchemy.orm import Session
+from fastapi import FastAPI
 
-from fast_zero.database import get_session
-from fast_zero.models import User
-from fast_zero.schemas import Message, Token, UserList, UserPublic, UserSchema
-from fast_zero.security import (
-    create_access_token,
-    get_current_user,
-    get_password_hash,
-    verify_password,
-)
+from fast_zero.routers import auth, users
 
 app = FastAPI()
+
+app.include_router(users.router)
+app.include_router(auth.router)
 
 
 @app.get('/')
 def read_root():
-    return {'message': 'ola mundo'}
-
-
-@app.post('/users/', status_code=201, response_model=UserPublic)
-def create_user(user: UserSchema, session: Session = Depends(get_session)):
-    db_user = session.scalar(
-        select(User).where(User.email == user.email)
-    )  ## pesquisa se existe o usuário
-    if db_user:
-        raise HTTPException(
-            status_code=400,
-            detail='Username already registered',  ## caso db_user retorne algo
-        )
-
-    hashed_password = get_password_hash(user.password)
-
-    db_user = User(  ## senão existir, cria um novo usuário
-        username=user.username, password=hashed_password, email=user.email
-    )
-    session.add(db_user)
-    session.commit()
-    session.refresh(db_user)
-    return db_user
-
-
-@app.get('/users/', response_model=UserList)
-def read_users(
-    skip: int = 0, limit: int = 100, session: Session = Depends(get_session)
-):
-    users = session.scalars(select(User).offset(skip).limit(limit)).all()
-    return {'users': users}
-
-
-@app.put('/users/{user_id}', response_model=UserPublic)
-def update_user(
-    user_id: int,
-    user: UserSchema,
-    session: Session = Depends(get_session),
-    current_user: User = Depends(get_current_user),
-):
-    if current_user.id != user_id:
-        raise HTTPException(status_code=400, detail='Not enough permissions')
-
-    current_user.username = user.username
-    current_user.password = user.password
-    current_user.email = user.email
-    session.commit()
-    session.refresh(current_user)
-
-    return current_user
-
-
-@app.delete('/users/{user_id}', response_model=Message)
-def delete_user(
-    user_id: int,
-    session: Session = Depends(get_session),
-    current_user: User = Depends(get_current_user),
-):
-    if current_user.id != user_id:
-        raise HTTPException(status_code=400, detail='Not enough permissions')
-    session.delete(current_user)
-    session.commit()
-
-    return {'detail': 'User deleted'}
-
-
-@app.post('/token', response_model=Token)
-def login_for_access_token(
-    form_data: OAuth2PasswordRequestForm = Depends(),
-    session: Session = Depends(get_session),
-):
-    user = session.scalar(select(User).where(User.email == form_data.username))
-    if not user:
-        raise HTTPException(
-            status_code=400, detail='Incorrect email or password'
-        )
-    if not verify_password(form_data.password, user.password):
-        raise HTTPException(
-            status_code=400, detail='Incorrect email or password'
-        )
-    access_token = create_access_token(data={'sub': user.email})
-
-    return {'access_token': access_token, 'token_type': 'bearer'}
+    return {'message': 'Olá mundo'}
