+++ tests/test_app.py
@@ -1,7 +1,6 @@
 from fastapi.testclient import TestClient
 
 from fast_zero.app import app
-from fast_zero.schemas import UserPublic
 
 client = TestClient(app)
 
@@ -9,75 +8,7 @@ client = TestClient(app)
 def test_root_deve_retornar_olamundo_e_200():
     response = client.get('/')
     assert response.status_code == 200
-    assert response.json() == {'message': 'ola mundo'}
-
-
-def test_create_user(client):
-    response = client.post(
-        '/users',
-        json={
-            'username': 'alice',
-            'email': 'alice@example.com',
-            'password': 'secret',
-        },
-    )
-    assert response.status_code == 201
-    assert response.json() == {
-        'username': 'alice',
-        'email': 'alice@example.com',
-        'id': 1,
-    }
-
-
-def test_create_user_already_created(client):
-    response = client.post(
-        '/users',
-        json={
-            'username': 'alice',
-            'email': 'alice@example.com',
-            'password': 'secret',
-        },
-    )
-    response = client.post(
-        '/users',
-        json={
-            'username': 'alice',
-            'email': 'alice@example.com',
-            'password': 'secret',
-        },
-    )
-    assert response.status_code == 400
-    assert response.json() == {'detail': 'Username already registered'}
-
-
-def test_read_users():
-    response = client.get('/users')
-    assert response.status_code == 200
-    assert response.json() == {'users': []}
-
-
-def test_read_users_with_users(client, user):
-    user_schema = UserPublic.model_validate(user).model_dump()
-    response = client.get('/users/')
-    assert response.json() == {'users': [user_schema]}
-
-
-def test_update_user(client, user, token):
-    response = client.put(
-        f'/users/{user.id}',
-        headers={'Authorization': f'Bearer {token}'},
-        json={
-            'username': 'bob',
-            'email': 'bob@example.com',
-            'password': 'mynewpassword',
-        },
-    )
-    assert response.status_code == 200
-    assert response.json() == {
-        'username': 'bob',
-        'email': 'bob@example.com',
-        'id': 1,
-    }
+    assert response.json() == {'message': 'Olá mundo'}
 
 
 """ def test_update_user_not_found(client, user):
@@ -91,34 +22,10 @@ def test_update_user(client, user, token):
     )
     assert response.status_code == 404
     assert response.json() == {'detail': 'User not found'}
-"""
 
-
-def test_delete_user(client, user, token):
-    response = client.delete(
-        f'/users/{user.id}',
-        headers={'Authorization': f'Bearer {token}'},
-    )
-    assert response.status_code == 200
-    assert response.json() == {'detail': 'User deleted'}
-
-
-"""
 def test_delete_user_not_found(client, user):
     response = client.delete('/users/0')
     assert response.status_code == 404
     assert response.json() == {'detail': 'User not found'}
 
 """
-
-
-def test_get_token(client, user):
-    response = client.post(
-        '/token',
-        data={'username': user.email, 'password': user.clean_password},
-    )
-    token = response.json()
-
-    assert response.status_code == 200
-    assert 'access_token' in token
-    assert 'token_type' in token
