+++ tests/test_security.py
@@ -1,11 +1,16 @@
 from jose import jwt
 
-from fast_zero.security import SECRET_KEY, create_access_token
+from fast_zero.security import create_access_token
+from fast_zero.settings import Settings
+
+settings = Settings()
 
 
 def test_jwt():
     data = {'test': 'test'}
     token = create_access_token(data)
-    decoded = jwt.decode(token, SECRET_KEY, algorithms=['HS256'])
+    decoded = jwt.decode(
+        token, settings.SECRET_KEY, algorithms=[settings.ALGORITHM]
+    )
     assert decoded['test'] == data['test']
     assert decoded['exp']
